Improve ANSIColors structure and expand color support
Rename ANSIColors module file name to be more fitting. Changed from 
'color_names' to 'ansi_colors'.
